# MS1/openapi.yaml
openapi: 3.0.3
info:
  title: MS1 Books API
  version: 1.0.0

servers:
  - url: http://localhost:3000

tags:
  - name: Books
  - name: Authors
  - name: Publishers

paths:

  # -----------------------------------------------------
  # BOOKS
  # -----------------------------------------------------
  /books:
    get:
      tags: [Books]
      summary: Alle Bücher abrufen (optional Suche mit ?q=)
      parameters:
        - in: query
          name: q
          description: Case-insensitive Suche in title/author/publisher
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Liste aller Bücher
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "500":
          description: Serverfehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [Books]
      summary: Neues Buch anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookCreate"
      responses:
        "201":
          description: Angelegtes Buch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Pflichtfelder fehlen oder ungültig
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    "500":
          description: Serverfehler
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /books/{id}:
    get:
      tags: [Books]
      summary: Einzelnes Buch abrufen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Buch gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Buch nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags: [Books]
      summary: Buch aktualisieren (Teilupdate)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookPatch"
      responses:
        "200":
          description: Buch aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Keine gültigen Felder übergeben
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Buch nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Books]
      summary: Buch löschen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Buch gelöscht
        "404":
          description: Buch nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # -----------------------------------------------------
  # AUTHORS
  # -----------------------------------------------------
  /authors:
    get:
      tags: [Authors]
      summary: Alle Autoren abrufen (optional Suche mit ?q=)
      parameters:
        - in: query
          name: q
          required: false
          description: Case-insensitive Suche in name
          schema:
            type: string
      responses:
        "200":
          description: Liste aller Autoren
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Author"

    post:
      tags: [Authors]
      summary: Neuen Autor anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorCreate"
      responses:
        "201":
          description: Autor angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /authors/{id}:
    get:
      tags: [Authors]
      summary: Einzelnen Autor abrufen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Autor gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "404":
          description: Autor nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags: [Authors]
      summary: Autor ändern
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorPatch"
      responses:
        "200":
          description: Autor aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Author"
        "400":
          description: Ungültiges Patch-Objekt
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Autor nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Authors]
      summary: Autor löschen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Autor gelöscht
        "404":
          description: Autor nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # -----------------------------------------------------
  # PUBLISHERS
  # -----------------------------------------------------
  /publishers:
    get:
      tags: [Publishers]
      summary: Alle Verlage abrufen (optional Suche mit ?q=)
      parameters:
        - in: query
          name: q
          required: false
          description: Case-insensitive Suche in name
          schema:
            type: string
      responses:
        "200":
          description: Liste aller Verlage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Publisher"

    post:
      tags: [Publishers]
      summary: Neuen Verlag anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherCreate"
      responses:
        "201":
          description: Verlag angelegt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /publishers/{id}:
    get:
      tags: [Publishers]
      summary: Einzelnen Verlag abrufen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verlag gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        "404":
          description: Verlag nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      tags: [Publishers]
      summary: Verlag ändern
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherPatch"
      responses:
        "200":
          description: Verlag aktualisiert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publisher"
        "400":
          description: Ungültiges Patch-Objekt
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Verlag nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [Publishers]
      summary: Verlag löschen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Verlag gelöscht
        "404":
          description: Verlag nicht gefunden
          headers:
            X-Fehlermeldung:
              $ref: "#/components/headers/X-Fehlermeldung"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# -----------------------------------------------------
# COMPONENTS
# -----------------------------------------------------
components:

  headers:
    X-Fehlermeldung:
      description: Beschreibung der Fehlerursache
      schema:
        type: string

  schemas:

    Book:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        author: { type: string }
        publisher: { type: string }
      required: [id, title, author, publisher]

    BookCreate:
      type: object
      properties:
        title: { type: string }
        author: { type: string }
        publisher: { type: string }
      required: [title, author, publisher]

    BookPatch:
      type: object
      minProperties: 1
      properties:
        title: { type: string }
        author: { type: string }
        publisher: { type: string }
      additionalProperties: false

    Author:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]

    AuthorCreate:
      type: object
      properties:
        name: { type: string }
      required: [name]

    AuthorPatch:
      type: object
      minProperties: 1
      properties:
        name: { type: string }
      additionalProperties: false

    Publisher:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
      required: [id, name]

    PublisherCreate:
      type: object
      properties:
        name: { type: string }
      required: [name]

    PublisherPatch:
      type: object
      minProperties: 1
      properties:
        name: { type: string }
      additionalProperties: false

    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]